product/requirements.txt
pandas
gcsfs
pyarrow
google-cloud-bigquery
google-cloud-aiplatform
